requirements.txt
cryptography==45.0.7
pillow==11.3.0
playsound==1.2.2
